requirements.txt
langchain
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-google-genai
langchain-text-splitters
streamlit
google-genai
fastembed
vaderSentiment
langdetect
deep-translator
python-dotenv
arxiv
plotly
wordcloud
pillow
pandas
numpy
pytest
pytest-mock
